app/requirements.txt
requests
pandas
numpy
flask
gunicorn
gspread
google-auth
python-dotenv
pytz
